protect create note page with auth and proper directing
